app/requirements.txt
fastapi==0.104.0
numpy==1.22.4
Pillow==10.0.0
rasterio==1.3.8
uvicorn==0.23.2
